+++ core/constants.py
@@ -7,6 +7,21 @@ BACKLINK_QUERYSTRING_NAME = 'return-link'
 # Define a strict subset of rich-text features only includes linebreaks
 RICHTEXT_FEATURES__MINIMAL = ()
 
+# Define rich-text features that enable the off-by-default blockquote element
+RICHTEXT_FEATURES__ARTICLE = (
+    'h2',
+    'h3',
+    'h4',
+    # 'h5', 'h6'  # also available if we want
+    'bold',
+    'italic',
+    'ol',
+    'ul',
+    'hr',
+    'link',
+    'blockquote',
+)
+
 AWS_S3_MAIN_HOSTNAME_OPTIONS = [
     # https://docs.aws.amazon.com/general/latest/gr/s3.html
     's3.amazonaws.com',  # most likely